Struct stm32l4x6_hal::timer::Timer

pub struct Timer<TIM> { /* fields omitted */ }

HW Timer

Methods

impl Timer<SYST>

pub fn syst<T: Into<Hertz>>(syst: SYST, timeout: T, clocks: Clocks) -> Self

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

impl Timer<TIM1>

pub fn tim1<T: Into<Hertz>>(
tim: TIM1,
timeout: T,
clocks: Clocks,
apb: &mut APB2
) -> Timer<TIM1>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM1

Paused timer and releases the TIM peripheral

impl Timer<TIM8>

pub fn tim8<T: Into<Hertz>>(
tim: TIM8,
timeout: T,
clocks: Clocks,
apb: &mut APB2
) -> Timer<TIM8>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM8

Paused timer and releases the TIM peripheral

impl Timer<TIM2>

pub fn tim2<T: Into<Hertz>>(
tim: TIM2,
timeout: T,
clocks: Clocks,
apb: &mut APB1
) -> Timer<TIM2>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM2

Paused timer and releases the TIM peripheral

impl Timer<TIM3>

pub fn tim3<T: Into<Hertz>>(
tim: TIM3,
timeout: T,
clocks: Clocks,
apb: &mut APB1
) -> Timer<TIM3>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM3

Paused timer and releases the TIM peripheral

impl Timer<TIM4>

pub fn tim4<T: Into<Hertz>>(
tim: TIM4,
timeout: T,
clocks: Clocks,
apb: &mut APB1
) -> Timer<TIM4>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM4

Paused timer and releases the TIM peripheral

impl Timer<TIM5>

pub fn tim5<T: Into<Hertz>>(
tim: TIM5,
timeout: T,
clocks: Clocks,
apb: &mut APB1
) -> Timer<TIM5>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM5

Paused timer and releases the TIM peripheral

impl Timer<TIM15>

pub fn tim15<T: Into<Hertz>>(
tim: TIM15,
timeout: T,
clocks: Clocks,
apb: &mut APB2
) -> Timer<TIM15>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM15

Paused timer and releases the TIM peripheral

impl Timer<TIM16>

pub fn tim16<T: Into<Hertz>>(
tim: TIM16,
timeout: T,
clocks: Clocks,
apb: &mut APB2
) -> Timer<TIM16>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM16

Paused timer and releases the TIM peripheral

impl Timer<TIM17>

pub fn tim17<T: Into<Hertz>>(
tim: TIM17,
timeout: T,
clocks: Clocks,
apb: &mut APB2
) -> Timer<TIM17>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM17

Paused timer and releases the TIM peripheral

impl Timer<TIM6>

pub fn tim6<T: Into<Hertz>>(
tim: TIM6,
timeout: T,
clocks: Clocks,
apb: &mut APB1
) -> Timer<TIM6>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM6

Paused timer and releases the TIM peripheral

impl Timer<TIM7>

pub fn tim7<T: Into<Hertz>>(
tim: TIM7,
timeout: T,
clocks: Clocks,
apb: &mut APB1
) -> Timer<TIM7>

Creates new instance of timer.

pub fn subscribe(&mut self, event: Event)

Starts listening for an event

pub fn unsubscribe(&mut self, event: Event)

Stops listening for an event

pub fn reset_overflow(&mut self)

Resets SR's UIF register to clear status of overflow.

Unless reset is done, Interrupt handler is going to be continiously called.

pub fn free(self) -> TIM7

Paused timer and releases the TIM peripheral